Fix inline value parsing
Edge case has a value that contains an inline symbol such as `;` or `#`.

Currently parsing a value such as:

```
value = val#ue
```

would result into just `val` if `IgnoreInlineComment` is false.

If `IgnoreInlineComment` is true it would result into the full value, but it's incorrect to depend on this.

With the fix value would be parsed as `val#ue`.